model api for testing and eventually usage
