+++ Bayesian/VAE/Models/model_api.h
@@ -1,0 +1,7 @@
+#include "Matrix.h"
+#include "model.h"
+#include "LinearModel.h"
+#include <stdlib.h>
+#include <time.h>
+
+void run_trained_model(double *image);
